20160608-Version
Coop4SWEEEM (Cooperation For Sustainable WEEE Management) simulates the influence of incentives on consumers of Electrical and Electronic Equipment. Based on the incentive, consumers decide to deliver (ot not) the waste to the post-consumer program (PCP). At the same time, the PCP works thanks to the cooperation between related producer (who assemble or import the device) and distributor (who sell devices).